+++ src/contentful/__tests__/contentfulData.test.ts
@@ -22,6 +22,7 @@ import { navigationItems } from '../../../tests/data/contentful/navigationItem.j
 import { navigationItemsFr } from '../../../tests/data/contentful/navigationItemFr.js'
 import { taxonomies } from '../../../tests/data/contentful/taxonomy.js'
 import { terms } from '../../../tests/data/contentful/term.js'
+import { normalJsonKeys } from '../contentfulDataNormal.js'
 
 /* Mock fetch */
 
@@ -41,10 +42,8 @@ describe('getContentfulData()', () => {
     config.cms.prodHost = 'cdn.contentful.com'
     config.cms.env = 'master'
     config.wholeTypes = ['page']
-    config.partialTypes = [
-      'navigation',
-      'navigationItem'
-    ]
+    config.partialTypes = ['navigation', 'navigationItem']
+    normalJsonKeys.add('json')
   })
 
   afterEach(() => {
@@ -67,12 +66,13 @@ describe('getContentfulData()', () => {
     config.env.cache = false
     config.wholeTypes = []
     config.partialTypes = []
+    normalJsonKeys.clear()
     testResetStore()
     resetFilters()
   })
 
   it('should throw an error if key is empty', async () => {
-    await expect(async () => await getContentfulData('')).rejects.toThrowError('No key')
+    await expect(async () => await getContentfulData('')).rejects.toThrow('No key')
   })
 
   it('should throw an error if no config credentials', async () => {
@@ -82,19 +82,19 @@ describe('getContentfulData()', () => {
     config.cms.prodCredential = ''
     config.cms.prodHost = ''
 
-    await expect(async () => await getContentfulData('key_2')).rejects.toThrowError('No credentials')
+    await expect(async () => await getContentfulData('key_2')).rejects.toThrow('No credentials')
   })
 
   it('should throw an error if invalid credentials', async () => {
     config.cms.devCredential = 'xyz'
 
-    await expect(async () => await getContentfulData('key_4')).rejects.toThrowError(mockFetchErrorMessage.auth)
+    await expect(async () => await getContentfulData('key_4')).rejects.toThrow(mockFetchErrorMessage.auth)
   })
 
   it('should throw an error if content type does not exist', async () => {
     await expect(async () => await getContentfulData('key_5', {
       content_type: 'none'
-    })).rejects.toThrowError(mockFetchErrorMessage.contentType)
+    })).rejects.toThrow(mockFetchErrorMessage.contentType)
   })
 
   it('should return empty array if items is empty', async () => {
@@ -196,7 +196,7 @@ describe('getContentfulData()', () => {
     await expect(async () => await getContentfulData('page_key_none', {
       content_type: 'page',
       'sys.id': 'none'
-    })).rejects.toThrowError('Bad fetch response')
+    })).rejects.toThrow('Bad fetch response')
   })
 
   it('should return array of taxonomies', async () => {
@@ -238,11 +238,9 @@ describe('getAllContentfulData()', () => {
     config.cms.prodCredential = 'lipsum'
     config.cms.prodHost = 'cdn.contentful.com'
     config.cms.env = 'master'
-    config.wholeTypes = ['page']
-    config.partialTypes = [
-      'navigation',
-      'navigationItem'
-    ]
+    config.wholeTypes = ['page', 'post']
+    config.partialTypes = ['navigation', 'navigationItem']
+    normalJsonKeys.add('json')
   })
 
   afterEach(() => {
@@ -263,6 +261,7 @@ describe('getAllContentfulData()', () => {
 
     config.wholeTypes = []
     config.partialTypes = []
+    normalJsonKeys.clear()
     testResetStore()
     resetFilters()
   })
@@ -270,7 +269,7 @@ describe('getAllContentfulData()', () => {
   it('should return throw an error if content type does not exist', async () => {
     config.wholeTypes = ['none']
 
-    await expect(async () => await getAllContentfulData()).rejects.toThrowError(mockFetchErrorMessage.contentType)
+    await expect(async () => await getAllContentfulData()).rejects.toThrow(mockFetchErrorMessage.contentType)
   })
 
   it('should return navigation and page data in all locales', async () => {
@@ -288,7 +287,8 @@ describe('getAllContentfulData()', () => {
         page: [
           ...pages.items,
           ...pagesFr.items
-        ]
+        ],
+        post: []
       }
     }
 
@@ -413,7 +413,8 @@ describe('getAllContentfulData()', () => {
             ...page,
             test: 'test'
           }
-        })
+        }),
+        post: []
       }
     }
 
@@ -452,6 +453,7 @@ describe('getAllContentfulData()', () => {
           ...pages.items,
           ...pagesFr.items
         ],
+        post: [],
         test: []
       }
     }