+++ src/contentful/contentfulDataNormal.ts
@@ -20,6 +20,13 @@ import { isNumber } from '../utils/number/number.js'
 import { config } from '../config/config.js'
 
 /**
+ * JSON field keys to skip normalizing.
+ *
+ * @type {Set<string>}
+ */
+const normalJsonKeys: Set<string> = new Set()
+
+/**
  * HTML tag from type.
  *
  * @private
@@ -224,6 +231,7 @@ const normalizeItem = (item: ContentfulDataItem, data: RenderItem[], isInternalL
   /* Type */
 
   let type = ''
+  let normalType = ''
 
   if (isString(itemCopy.sys?.type)) {
     type = itemCopy.sys.type
@@ -234,10 +242,23 @@ const normalizeItem = (item: ContentfulDataItem, data: RenderItem[], isInternalL
   }
 
   if (type && type !== 'Link') {
-    newItem.contentType = normalizeContentType(type)
+    normalType = normalizeContentType(type)
+    newItem.contentType = normalType
 
     if (isString(config.renderTypes[type])) {
       newItem.renderType = config.renderTypes[type]
+    }
+  }
+
+  /* Date */
+
+  if (config.wholeTypes.includes(normalType)) {
+    if (isStringStrict(itemCopy.sys?.createdAt)) {
+      newItem.createdAt = itemCopy.sys.createdAt
+    }
+
+    if (isStringStrict(itemCopy.sys?.updatedAt)) {
+      newItem.updatedAt = itemCopy.sys.updatedAt
     }
   }
 
@@ -283,7 +304,7 @@ const normalizeItem = (item: ContentfulDataItem, data: RenderItem[], isInternalL
             })
           }
         } else {
-          newItem[prop] = normalizeItem(field, data, prop === 'internalLink')
+          newItem[prop] = normalJsonKeys.has(prop) ? field : normalizeItem(field, data, prop === 'internalLink' || isInternalLink)
         }
       } else if (isArrayStrict(field)) {
         newItem[prop] = field.map(f => {
@@ -351,4 +372,7 @@ const normalizeContentfulData = (data: ContentfulDataItem[], _newData: RenderIte
 
 /* Exports */
 
-export { normalizeContentfulData }
+export {
+  normalizeContentfulData,
+  normalJsonKeys
+}